+++ changelog.md
@@ -1,4 +1,13 @@
+0.0.2
+=====
+
+* Download TV show subtitle files.
+
+* Filter movie files, so you can drag any kind of file, but only movie files are take into account.
+
 0.0.1
 =====
 
 * Initial stage: frontend, install packages and some assets by default.
+
+* Download movie subtitle files. Not very accurate.